feat(Game): added charge animation networked

+++ src/POCEngine/System/SpawnMob.cpp
@@ -33,8 +33,8 @@ void SpawnMob::update(GameEngine::ComponentsContainer &componentsContainer, Game
         int tick = currentMapContent.getInt("/mobs/" + std::to_string(i) + "/tick");
 
         if (currentTick == tick) {
-            int posX = currentMapContent.getInt("/mobs/" + std::to_string(i) + "/position/x");
-            int posY = currentMapContent.getInt("/mobs/" + std::to_string(i) + "/position/y");
+            int posX = currentMapContent.getFloat("/mobs/" + std::to_string(i) + "/position/x");
+            int posY = currentMapContent.getFloat("/mobs/" + std::to_string(i) + "/position/y");
             Utils::Vect2 position(posX, posY);
 
             bool dropPowerup = currentMapContent.getBool("/mobs/" + std::to_string(i) + "/dropPowerup");
